# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/role": {
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Роль успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseWithId"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Создание новой роли",
        "tags": [
          "role"
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "страница",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "отображаемых записей на странице",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "порядок сортировки ASC/DESC",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "поиск",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "description": "фильтры (объект в виде JSON-строки)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список ролей успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение списка всех ролей",
        "tags": [
          "role"
        ]
      }
    },
    "/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID роли",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Роль успешно найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithUsersAndPermissions"
                }
              }
            }
          },
          "404": {
            "description": "Роль не найдена"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение роли по ID",
        "tags": [
          "role"
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID роли",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные роли успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithUsersAndPermissions"
                }
              }
            }
          },
          "404": {
            "description": "Роль не найдена"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Обновление данных роли",
        "tags": [
          "role"
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID роли",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Роль успешно удалена"
          },
          "404": {
            "description": "Роль не найдена"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Удаление роли",
        "tags": [
          "role"
        ]
      }
    },
    "/role/set-permissions/{roleName}": {
      "post": {
        "operationId": "RoleController_setPermissionToRole",
        "parameters": [
          {
            "name": "roleName",
            "required": true,
            "in": "path",
            "description": "Название роли",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Массив названий разрешений",
                    "example": [
                      "can_read",
                      "can_write"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Разрешения успешно назначены роли",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithUsersAndPermissions"
                }
              }
            }
          },
          "404": {
            "description": "Роль или разрешения не найдены"
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Назначение разрешений роли",
        "tags": [
          "role"
        ]
      }
    },
    "/permission": {
      "post": {
        "operationId": "PermissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Разрешение успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Создание нового разрешения",
        "tags": [
          "permission"
        ]
      },
      "get": {
        "operationId": "PermissionController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "страница",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "отображаемых записей на странице",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "порядок сортировки ASC/DESC",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "поиск",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "description": "фильтры (объект в виде JSON-строки)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список разрешений успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение списка всех разрешений",
        "tags": [
          "permission"
        ]
      }
    },
    "/permission/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID разрешения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Разрешение успешно найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Разрешение не найдено"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение разрешения по ID",
        "tags": [
          "permission"
        ]
      },
      "patch": {
        "operationId": "PermissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID разрешения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные разрешения успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Разрешение не найдено"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Обновление данных разрешения",
        "tags": [
          "permission"
        ]
      },
      "delete": {
        "operationId": "PermissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID разрешения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Разрешение успешно удалено"
          },
          "404": {
            "description": "Разрешение не найдено"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Удаление разрешения",
        "tags": [
          "permission"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход в систему. Устанавливаются куки access_token и refresh_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        },
        "summary": "Вход в систему",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Токен успешно обновлен. Устанавливаются новые куки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Невалидный refresh token"
          }
        },
        "summary": "Обновление токена",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Успешный выход. Куки удалены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Выход из системы",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Passport API",
    "description": "API документация для системы управления пользователями, ролями и разрешениями\n\nСпецификация OpenAPI доступна в формате JSON: [openapi.json](/api/openapi.json)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "UserWithoutRoles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя (уникальное)"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя (уникальный)"
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "writeOnly": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Статус активности пользователя",
            "default": true
          },
          "isVerified": {
            "type": "boolean",
            "description": "Статус верификации пользователя",
            "default": false
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона пользователя"
          },
          "authProvider": {
            "type": "string",
            "description": "Провайдер аутентификации (например, local, google, github)"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя (например, en, ru)"
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс пользователя"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "password",
          "isActive",
          "isVerified"
        ]
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название роли",
            "example": "user",
            "minLength": 3
          }
        },
        "required": [
          "name"
        ]
      },
      "RoleResponseWithId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор роли"
          },
          "name": {
            "type": "string",
            "description": "Название роли (уникальное)"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "PermissionWithoutRoles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RoleWithUsersAndPermissions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "ID роли"
          },
          "name": {
            "type": "string",
            "example": "admin",
            "description": "Название роли"
          },
          "users": {
            "description": "Список пользователей с этой ролью",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserWithoutRoles"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionWithoutRoles"
            }
          }
        },
        "required": [
          "id",
          "name",
          "users",
          "permissions"
        ]
      },
      "RoleListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Список ролей",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleWithUsersAndPermissions"
            }
          },
          "total": {
            "type": "number",
            "example": 10,
            "description": "Общее количество записей"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название роли",
            "example": "user",
            "minLength": 3
          }
        }
      },
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название разрешения",
            "example": "can_read:users",
            "minLength": 3
          }
        },
        "required": [
          "name"
        ]
      },
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roles": {
            "description": "Роли, связанные с этим разрешением",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponseWithId"
            }
          }
        },
        "required": [
          "id",
          "name",
          "roles"
        ]
      },
      "PermissionListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Список разрешений",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponse"
            }
          },
          "total": {
            "type": "number",
            "description": "Общее количество записей"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название разрешения",
            "example": "can_read:users",
            "minLength": 3
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "admin"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение о результате операции"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}